# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с обработчиком изображений",
        "title": "Image Processor Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "RidusM",
            "email": "stormkillpeople@gmail.com"
        },
        "license": {
            "name": "MIT-0",
            "url": "https://github.com/aws/mit-0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Проверка доступности сервиса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Сервис доступен",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/image/{id}": {
            "get": {
                "description": "Возвращает обработанное изображение или статус задачи",
                "produces": [
                    "application/json",
                    "image/jpeg",
                    "image/png"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Получить изображение или статус",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "processed",
                        "description": "Версия: processed|original|thumb",
                        "name": "version",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обработанное изображение",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "202": {
                        "description": "Изображение ещё обрабатывается",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImageStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Изображение не найдено",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет изображение и все его версии (оригинал, обработанное, миниатюру)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Удалить изображение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Изображение удалено"
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/image/{id}/status": {
            "get": {
                "description": "Возвращает текущий статус задачи обработки изображения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Получить статус обработки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обработка завершена",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImageStatusResponse"
                        }
                    },
                    "202": {
                        "description": "Обработка в процессе",
                        "schema": {
                            "$ref": "#/definitions/handlers.ImageStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Загружает изображение на обработку (ресайз, миниатюра, водяной знак)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "Загрузить изображение",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Файл изображения",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON с опциями обработки",
                        "name": "options",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Изображение принято в обработку",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Файл слишком большой",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "details": {
                    "type": "string",
                    "example": "image with id abc123 does not exist"
                },
                "error": {
                    "type": "string",
                    "example": "image not found"
                }
            }
        },
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "time": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                }
            }
        },
        "handlers.ImageStatusResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "error": {
                    "type": "string",
                    "example": "resize failed"
                },
                "filename": {
                    "type": "string",
                    "example": "photo.jpg"
                },
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6"
                },
                "progress": {
                    "type": "integer",
                    "example": 45
                },
                "status": {
                    "type": "string",
                    "example": "processing"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:05Z"
                }
            }
        },
        "handlers.UploadResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "photo.jpg"
                },
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6"
                },
                "message": {
                    "type": "string",
                    "example": "Image uploaded successfully"
                },
                "size": {
                    "type": "integer",
                    "example": 2048576
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        }
    }
}